TD-3896-Controller refactor - DataService reference removed and service classes added accordingly.
